Split Build Command into two commands
