+++ packages/sfpowerscripts-cli/messages/quickbuild.json
@@ -1,5 +1,5 @@
 {
-    "commandDescription": "Build all packages (unlocked/source/data) in a repo in parallel, respecting the dependency of each packages and generate artifacts to a provided directory",
+    "commandDescription": "Build all packages (unlocked/source/data) in a repo in parallel, respecting the dependency of each packages and generate artifacts to a provided directory without validating individual dependencies",
     "devhubAliasFlagDescription": "Provide the alias of the devhub previously authenticated, default value is HubOrg if using the Authenticate Devhub task",
     "diffCheckFlagDescription": "Only build the packages which have changed by analyzing previous tags",
     "gitTagFlagDescription": "Tag the current commit ID with an annotated tag containing the package name and version - does not push tag",
@@ -7,8 +7,7 @@
     "buildNumberFlagDescription": "The build number to be used for source packages, Unlocked Packages will be assigned the buildnumber from Saleforce directly if using .NEXT",
     "configFilePathFlagDescription": "Path in the current project directory containing  config file for the packaging org",
     "projectDirectoryFlagDescription": "The project directory should contain a sfdx-project.json for this command to succeed",
-    "artifactDirectoryFlagDescription": "The directory where the generated artifact is to be written",
-    "isValidationToBeSkippedFlagDescription": "Skips validation of dependencies, package ancestors, and metadata during package version creation. Skipping validation reduces the time it takes to create a new package version, but package versions created without validation can’t be promoted.",
+    "artifactDirectoryFlagDescription": "The directory where the artifact is to be written",
     "waitTimeFlagDescription": "Wait time for command to finish in minutes",
     "executorCountFlagDescription": "Number of parallel package task schedulors",
     "branchFlagDescription":"The git branch that this build is triggered on, Useful for metrics and general identification purposes",
