+++ packages/sfpowerscripts-cli/package.json
@@ -53,7 +53,7 @@
         "description": "Orchestrate packages from a monorepo through its lifecycle, driven by descriptors in your sfdx-project.json"
       },
       "sfpowerscripts:pool": {
-        "description": "Manage the pools crated by the orchestrator"
+        "description": "Manage the pools created by the orchestrator"
       }
     },
     "devPlugins": [
