+++ packages/sfpowerscripts-cli/src/impl/Stage.ts
@@ -3,6 +3,7 @@ export enum Stage
   ANALYZE="analyze",
   PREPARE="prepare",
   VALIDATE="validate",
+  QUICKBUILD="quickbuild",
   BUILD="build",
   DEPLOY="deploy",
   PROMOTE="promote",
